feat(LadderGameInformation): 참가자 수와 실행결과의 수가 맞지 않으면, throw exception
