+++ README.md
@@ -47,7 +47,6 @@ pobi               cron               jh
 - all을 입력 받으면, 실행결과 모두를 출력한다.
 
 ### TODO
-- 참가자 수와 실행결과의 수가 맞지 않으면, throw exception
 ### DONE
 - 가로 사다리를 생성
     1. 랜덤하게 생성할지 말지를 결정하고, 이전 가로 사다리가 생성되어 있으면 생성하지 않음
@@ -78,3 +77,4 @@ pobi               cron               jh
 - 결과 출력 버그 수정
 - 입력 받은 참가자 명에 따른 결과를 출력한다.
 - all을 입력 받으면, 실행결과 모두를 출력한다.
+- 참가자 수와 실행결과의 수가 맞지 않으면, throw exception
