--- src/test/java/ladder/domain/LadderGameResultTest.java
@@ -1,10 +1,0 @@
-package ladder.domain;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class LadderGameResultTest {
-
-}
