+++ src/test/java/ladder/domain/LadderGameTest.java
