+++ src/test/java/ladder/domain/LadderSizeTest.java
